requirements.txt
# Core ML/DL Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
torchaudio>=2.0.0
tensorflow>=2.13.0  # Alternative to PyTorch

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
pydub>=0.25.1

# Feature Extraction
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Model Training & Tracking
tensorboard>=2.13.0
wandb>=0.15.0
tqdm>=4.65.0

# Configuration & Utils
pyyaml>=6.0
python-dotenv>=1.0.0
hydra-core>=1.3.0

# Data Handling
kaggle>=1.5.16

# Jupyter & Visualization
jupyter>=1.0.0
ipython>=8.14.0
plotly>=5.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
